+++ skidl/Alias.py
@@ -33,8 +33,6 @@ from builtins import str
 
 from future import standard_library
 
-from .utilities import *
-
 standard_library.install_aliases()