+++ skidl/__init__.py
@@ -17,5 +17,25 @@ The script can also
 check the resulting circuitry for electrical rule violations.
 """
 
+from . import tools
+from .Alias import *
+from .Bus import *
+from .Circuit import *
+from .defines import *
+from .Interface import *
+from .logger import *
+from .Net import *
+from .netclass import *
 from .netlist_to_skidl import *
+from .NetPinList import *
+from .Network import *
+from .Note import *
+from .Package import *
+from .Part import *
+from .part_query import *
+from .Pin import *
+from .ProtoNet import *
+from .py_2_3 import *
+from .SchLib import *
 from .skidl import *
+from .utilities import *