+++ tests/test_package.py
@@ -22,20 +22,25 @@ def test_package_1():
 
         vin += r1[1], c1[1]  # Connect the input to the first resistor.
         gnd += r2[2], c2[2]  # Connect the second resistor to ground.
-        vout += r1[2], c1[2], r2[1], c2[1]   # Output comes from the connection of the two resistors.
+        vout += (
+            r1[2],
+            c1[2],
+            r2[1],
+            c2[1],
+        )  # Output comes from the connection of the two resistors.
 
     resdiv1 = resdiv()
     resdiv2 = resdiv()
 
-    vin, vout, gnd = Net('VI'), Net('VO'), Net('GND')
+    vin, vout, gnd = Net("VI"), Net("VO"), Net("GND")
 
     resdiv1.gnd += gnd
     resdiv1.vin += vin
-    resdiv1.vout += vout 
+    resdiv1.vout += vout
 
-    resdiv2['gnd'] += gnd
-    resdiv2['vin'] += vin
-    resdiv2['vout'] += vout 
+    resdiv2["gnd"] += gnd
+    resdiv2["vin"] += vin
+    resdiv2["vout"] += vout
 
     gnd += Pin()
     vin += Pin()
@@ -50,22 +55,22 @@ def test_package_1():
     assert len(resdiv1.vin) == 5
     assert len(resdiv1.vout) == 8
 
-    assert len(resdiv2['gnd']) == 5
-    assert len(resdiv2['vin']) == 5
-    assert len(resdiv2['vout']) == 8
+    assert len(resdiv2["gnd"]) == 5
+    assert len(resdiv2["vin"]) == 5
+    assert len(resdiv2["vout"]) == 8
+
 
 def test_package_2():
     """Test nested packages and interconnection."""
 
     @package
     def rc_rc(gnd, vin, vout):
-
         @package
         def rc(gnd, vin, vout):
-            r = Part('Device', 'R')
-            c = Part('Device', 'C')
+            r = Part("Device", "R")
+            c = Part("Device", "C")
             vin & r & vout & c & gnd
-        
+
         stage1 = rc()
         stage2 = rc()
 
@@ -79,11 +84,11 @@ def test_package_2():
     rc_rc_1 = rc_rc()
     rc_rc_2 = rc_rc()
 
-    rc_rc_1.vin += Net('VI')
+    rc_rc_1.vin += Net("VI")
     rc_rc_1.vout += Net()
     rc_rc_2.vin += rc_rc_1.vout
-    rc_rc_2.vout += Net('VO')
-    rc_rc_1.gnd += Net('GND')
+    rc_rc_2.vout += Net("VO")
+    rc_rc_1.gnd += Net("GND")
     rc_rc_2.gnd += rc_rc_1.gnd
 
     default_circuit.instantiate_packages()
@@ -96,6 +101,6 @@ def test_package_2():
     assert len(rc_rc_1.vin) == 1
     assert len(rc_rc_1.vout) == 3
 
-    assert len(rc_rc_2['gnd']) == 4
-    assert len(rc_rc_2['vin']) == 3
-    assert len(rc_rc_2['vout']) == 2
+    assert len(rc_rc_2["gnd"]) == 4
+    assert len(rc_rc_2["vin"]) == 3
+    assert len(rc_rc_2["vout"]) == 2